Add test case for DataFrame.from_dict
