+++ pandas/tests/frame/test_constructors.py
@@ -5,11 +5,7 @@ from collections import (
     defaultdict,
     namedtuple,
 )
-from collections.abc import (
-    Iterator,
-    Mapping,
-    Sequence,
-)
+from collections.abc import Iterator
 from dataclasses import make_dataclass
 from datetime import (
     date,
@@ -65,6 +61,10 @@ from pandas.arrays import (
     SparseArray,
     TimedeltaArray,
 )
+from pandas.tests.frame.common import (
+    DictWrapper,
+    ListWrapper,
+)
 
 MIXED_FLOAT_DTYPES = ["float16", "float32", "float64"]
 MIXED_INT_DTYPES = [
@@ -77,35 +77,6 @@ MIXED_INT_DTYPES = [
     "int32",
     "int64",
 ]
-
-
-class DictWrapper(Mapping):
-    _dict: dict
-
-    def __init__(self, d: dict) -> None:
-        self._dict = d
-
-    def __getitem__(self, key):
-        return self._dict[key]
-
-    def __iter__(self):
-        return self._dict.__iter__()
-
-    def __len__(self):
-        return self._dict.__len__()
-
-
-class ListWrapper(Sequence):
-    _list: list
-
-    def __init__(self, lst: list) -> None:
-        self._list = lst
-
-    def __getitem__(self, i):
-        return self._list[i]
-
-    def __len__(self):
-        return self._list.__len__()
 
 
 class TestDataFrameConstructors:
@@ -2950,7 +2921,7 @@ class TestDataFrameConstructorWithDatetimeTZ:
         tm.assert_series_equal(df["A"], Series(idx, name="A"))
         tm.assert_series_equal(df["B"], Series(dr, name="B"))
 
-    def test_from_dict_with_mapping(self):
+    def test_from_mapping(self):
         idx = Index(date_range("20130101", periods=3, tz="US/Eastern"), name="foo")
         dr = date_range("20130110", periods=3)
